A new upload input and fixes

+++ src/cake/app/plugins/jj_users/models/user_profile.php
@@ -32,39 +32,56 @@ class UserProfile extends JjUsersAppModel
 			'notempty' => array('rule' => array('notempty')),
 		)
 	);
-
+	
 	var $hasAndBelongsToMany = array(
 		'UserPermission' => array(
 			'className' => 'JjUsers.UserPermission', 
 			'joinTable' => 'user_profiles_user_permissions',
+		),
+		'UserUser' => array(
+			'className' => 'JjUsers.UserUser',
+			'joinTable' => 'user_users_user_profiles',
+			'unique' => false
 		)
 	);
-	
-	
+
 	function backDelete($id)
 	{
-		$this->bindModel(array(
-			'hasMany' => array(
-				'UserUsersUserProfile' => array(
-					'className' => 'jjUsers.UserUsersUserProfile',
-				)
-			)
-		));
-		
-		$this->UserUsersUserProfile->deleteAll(array('user_profile_id' => $id), false);
+		$this->contain('UserUser');
+		$data = $this->findById($id);
+		if (!empty($data['UserUser']))
+			return false;
 		return $this->delete($id);
 	}
 	
-	/* Creates a blank row in the table. It is part of the backstage contract.
-	 *
-	 */
+/**
+ * Creates a blank row in the table. It is part of the backstage contract.
+ */
 	function createEmpty()
 	{
-		
-		$data = $this->saveAll(array($this->alias => array()), array('validate' => false));
-		$data = $this->find('first', array('conditions' => array($this->alias.'.id' => $this->id)));
-		
-		return $data;
+		$saved = $this->saveAll(array($this->alias => array()), array('validate' => false));
+		if ($saved)
+			return $this->id;
+		return false;
 	}
-	
+
+/**
+ * Ensures that all related users be updated forcing permission reload of logged users.
+ * 
+ * @access public
+ */
+	function afterSave()
+	{
+		$this->contain('UserUser');
+		$data = $this->read();
+
+		$usersId = Set::extract('/UserUser/id', $data);
+		if (!empty($usersId))
+		{
+			$this->UserUser->updateAll(
+				array('UserUser.modified' => '"' . date('Y-m-d H:i:s') . '"'),
+				array('UserUser.id' => $usersId)
+			);
+		}
+	}
 }
